Reapply "Initial move of SingleColumnTransformer to  "core", and associated refactoring"
This reverts commit 90c8191602b87c95a26510aaecfc2e6af4ac8009.
