+++ skrub/_clean_categories.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from . import _dataframe as sbd
 from ._dispatch import dispatch, raise_dispatch_unregistered_type
-from ._single_column_transformer import RejectColumn, SingleColumnTransformer
+from .core._single_column_transformer import RejectColumn, SingleColumnTransformer
 
 __all__ = ["CleanCategories"]
