+++ skrub/_datetime_encoder.py
@@ -7,8 +7,8 @@ from sklearn.utils.validation import check_is_fitted
 
 from . import _dataframe as sbd
 from ._dispatch import dispatch
-from ._single_column_transformer import RejectColumn, SingleColumnTransformer
 from ._sklearn_compat import TransformerTags
+from .core._single_column_transformer import RejectColumn, SingleColumnTransformer
 
 __all__ = ["DatetimeEncoder"]
