+++ skrub/_drop_uninformative.py
@@ -3,7 +3,7 @@ import numbers
 from sklearn.utils.validation import check_is_fitted
 
 from . import _dataframe as sbd
-from ._single_column_transformer import SingleColumnTransformer
+from .core._single_column_transformer import SingleColumnTransformer
 
 __all__ = ["DropUninformative"]
