+++ skrub/_gap_encoder.py
@@ -19,8 +19,8 @@ from sklearn.utils.extmath import row_norms, safe_sparse_dot
 from sklearn.utils.validation import _num_samples, check_is_fitted
 
 from . import _dataframe as sbd
-from ._single_column_transformer import RejectColumn, SingleColumnTransformer
 from ._utils import unique_strings
+from .core._single_column_transformer import RejectColumn, SingleColumnTransformer
 
 
 class GapEncoder(TransformerMixin, SingleColumnTransformer):
