+++ skrub/_minhash_encoder.py
@@ -13,9 +13,9 @@ from sklearn.utils.validation import check_is_fitted
 
 from . import _dataframe as sbd
 from ._fast_hash import ngram_min_hash
-from ._single_column_transformer import RejectColumn, SingleColumnTransformer
 from ._string_distances import get_unique_ngrams
 from ._utils import LRUDict, unique_strings
+from .core._single_column_transformer import RejectColumn, SingleColumnTransformer
 
 NoneType = type(None)
