+++ skrub/_text_encoder.py
@@ -9,9 +9,9 @@ from sklearn.utils.validation import check_is_fitted
 
 from . import _dataframe as sbd
 from ._scaling_factor import scaling_factor
-from ._single_column_transformer import SingleColumnTransformer
 from ._to_str import ToStr
 from ._utils import import_optional_dependency, unique_strings
+from .core._single_column_transformer import SingleColumnTransformer
 from .datasets._utils import get_data_dir
