+++ skrub/_to_categorical.py
@@ -1,5 +1,5 @@
 from . import _dataframe as sbd
-from ._single_column_transformer import RejectColumn, SingleColumnTransformer
+from .core._single_column_transformer import RejectColumn, SingleColumnTransformer
 
 __all__ = ["ToCategorical"]
