+++ skrub/core/_single_column_transformer.py
@@ -7,7 +7,7 @@ import textwrap
 from sklearn.base import BaseEstimator
 from sklearn.utils.validation import check_is_fitted
 
-from . import _dataframe as sbd
+from .. import _dataframe as sbd
 
 __all__ = ["SingleColumnTransformer", "RejectColumn"]
