+++ skrub/tests/test_clean_null_strings.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from skrub._clean_null_strings import CleanNullStrings, _trim_whitespace_only
-from skrub._single_column_transformer import RejectColumn
+from skrub.core._single_column_transformer import RejectColumn
 
 
 def test_clean_null_strings(df_module):
