+++ skrub/tests/test_text_encoder.py
@@ -8,8 +8,8 @@ from sklearn.base import clone
 
 import skrub._dataframe as sbd
 from skrub import TableVectorizer, TextEncoder
-from skrub._single_column_transformer import RejectColumn
 from skrub._text_encoder import ModelNotFound
+from skrub.core._single_column_transformer import RejectColumn
 
 
 @pytest.fixture
