+++ skrub/tests/test_to_categorical.py
@@ -1,8 +1,8 @@
 import pytest
 
 from skrub import _dataframe as sbd
-from skrub._single_column_transformer import RejectColumn
 from skrub._to_categorical import ToCategorical
+from skrub.core._single_column_transformer import RejectColumn
 
 
 def test_to_categorical(df_module):
