+++ skrub/tests/test_to_float.py
@@ -2,11 +2,11 @@ import numpy as np
 import pytest
 
 from skrub import _dataframe as sbd
-from skrub._single_column_transformer import RejectColumn
 from skrub._to_categorical import ToCategorical
 from skrub._to_datetime import ToDatetime
 from skrub._to_float import ToFloat
 from skrub.conftest import skip_polars_installed_without_pyarrow
+from skrub.core._single_column_transformer import RejectColumn
 
 
 def is_float32(df_module, column):
