+++ skrub/tests/test_to_str.py
@@ -4,10 +4,10 @@ import pytest
 from packaging.version import parse
 
 from skrub import _dataframe as sbd
-from skrub._single_column_transformer import RejectColumn
 from skrub._to_datetime import ToDatetime
 from skrub._to_str import ToStr
 from skrub.conftest import skip_polars_installed_without_pyarrow
+from skrub.core._single_column_transformer import RejectColumn
 
 
 def test_to_str(df_module):
